@@ -13,8 +13,9 @@
     "@types/uuid": "^8.3.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "3.4.1",
-    "typescript": "~3.7.2",
+    "typescript": "^4.5.5",
     "uuid": "^8.3.2"
   },
   "homepage": "https://dezo103.github.io/HW",
