@@ -1,17 +1,18 @@
 import React from 'react'
 import Header from './Header'
-import Routes from './Routes'
+import AllRoutes from './Routes'
+import {BrowserRouter} from "react-router-dom";
 
 function HW5() {
     return (
         <div>
             {/*в gh-pages лучше работает HashRouter*/}
             {/*<HashRouter>*/}
+            <BrowserRouter>
+                <Header/>
 
-            <Header/>
-
-            <Routes/>
-
+                <AllRoutes/>
+            </BrowserRouter>
             {/*</HashRouter>*/}
         </div>
     )
